Refactor test_print and enable debug in more modules
Moved the test_print function to a new Printn namespace in src/print.php and updated all relevant files to use it. Added error reporting toggles to additional modules for easier debugging when 'test' is set in GET or COOKIE. Cleaned up require.php by removing the old test_print implementation.
