+++ public_html/new_html/check.php
@@ -1,4 +1,9 @@
 <?php
+if (isset($_GET['test']) || isset($_COOKIE['test'])) {
+    ini_set('display_errors', 1);
+    ini_set('display_startup_errors', 1);
+    error_reporting(E_ALL);
+}
 
 $revid = $_GET['revid'] ?? '';
