+++ public_html/new_html/main.php
@@ -1,13 +1,9 @@
 <?php
 header("Access-Control-Allow-Origin: *");
-if (isset($_GET['test']) || isset($_COOKIE['test'])) {
-    ini_set('display_errors', 1);
-    ini_set('display_startup_errors', 1);
-    error_reporting(E_ALL);
-}
 
 require_once __DIR__ . "/require.php";
 
+use function Printn\test_print;
 use function Wikitext\get_wikitext;
 use function Segments\html_to_seg;
 use function Html\wiki_text_to_html;
