+++ public_html/new_html/open.php
@@ -1,11 +1,5 @@
 <?php
 
-if (isset($_GET['test']) || isset($_COOKIE['test'])) {
-    ini_set('display_errors', 1);
-    ini_set('display_startup_errors', 1);
-    error_reporting(E_ALL);
-}
-
 require_once __DIR__ . "/require.php";
 
 use function HtmlFixes\remove_data_parsoid;
