+++ public_html/new_html/src/WikiText/fixes/expend_refs.php
@@ -9,6 +9,7 @@ use function Fixes\ExpendRefs\refs_expend_work;
 
 */
 
+use function Printn\test_print;
 use function WikiParse\Reg_Citations\get_full_refs;
 use function WikiParse\Reg_Citations\getShortCitations;
