+++ public_html/new_html/src/post.php
@@ -10,6 +10,8 @@ use function Post\post_url_params_result;
 // https://mdwiki.org/w/rest.php/v1/page/Sympathetic_crashing_acute_pulmonary_edema/html
 // https://mdwiki.org/w/rest.php/v1/revision/1420795/html
 
+use function Printn\test_print;
+
 $usr_agent = 'WikiProjectMed Translation Dashboard/1.0 (https://medwiki.toolforge.org/; tools.medwiki@toolforge.org)';
 
 function post_url_params_result(string $endPoint, array $params = []): string
