+++ public_html/new_html/src/print.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Printn;
+/*
+use function Printn\test_print;
+*/
+
+function test_print($str)
+{
+    if (isset($_GET['test']) || defined('DEBUGX')) {
+        echo $str;
+        echo "\n";
+    }
+}
